+++ packages/cli/cypress/e2e/index.cy.js
@@ -1,35 +1,39 @@
 const { isBindingElement } = require("typescript");
 
-describe("cypress tests", () => {
+describe("index page tests", () => {
   beforeEach(() => {
     cy.visit("/");
   });
 
-  it("should show the preview server index page", () => {
-    cy.get("h1").should("contain", "previews");
-    cy.get(".email-group").should("have.length", 2);
-
-    // TextEmail group should have 3 previews
-    cy.get(".email-group")
-      .contains("TextEmail")
-      .siblings(".email-container")
-      .should("have.length", 3);
-
-    // Welcome group should have 1 preview
-    cy.get(".email-group")
-      .contains("Welcome")
-      .siblings(".email-container")
-      .should("have.length", 1);
-  });
-
-  it("clicking a link should load the preview", () => {
-    cy.get(".email-group .email-container .email")
-      .contains("accountDeleted")
-      .click();
-
+  it("should redirect index to previewFunction with tree", () => {
     cy.location("pathname").should(
-      "match",
-      /previews\/TextEmail\/accountDeleted/
+      "eq",
+      "/previews/AccountCreated/accountCreated"
     );
+
+    cy.contains("accountCreated")
+      .should("have.attr", "aria-selected", "true")
+      .should("have.attr", "role", "treeitem");
+    cy.contains("Emails")
+      .should("have.attr", "aria-expanded", "true")
+      .should("have.attr", "aria-selected", "false")
+      .should("have.attr", "role", "treeitem");
+    cy.contains("Reservation")
+      .should("have.attr", "aria-expanded", "true")
+      .should("have.attr", "role", "treeitem")
+      .should("have.attr", "role", "treeitem");
+    cy.contains("ResetPassword")
+      .should("have.attr", "aria-expanded", "true")
+      .should("have.attr", "role", "treeitem");
+
+    cy.contains("reservationChanged").click();
+    cy.location("pathname").should(
+      "eq",
+      "/previews/Reservation/reservationChanged"
+    );
+
+    cy.contains("Compact view").click();
+
+    cy.get("[aria-selected=true]").contains("reservationChanged");
   });
 });